+++ process_monitor/sampler/main.cpp
@@ -14,6 +14,6 @@ int main() {
         } else {
             std::cout << "CPU Usage: " << usage << "%\n";
         }
-        std::this_thread::sleep_for(std::chrono::seconds(1));
+        std::this_thread::sleep_for(std::chrono::milliseconds(100));
     }
 }